websocket-server/pyproject.toml
[project]
name = "websocket-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "kafka-python>=2.2.15",
    "logger",
    "lz4>=4.4.4",
    "metrics",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
    "uvicorn[standard]>=0.37.0",
]

[tool.uv.sources]
logger = { path = "../common/logger" }
metrics = { path = "../common/metrics" }
